move getter specs to expressions module
